+++ CHANGELOG.md
@@ -13,6 +13,8 @@ CentOS-6 6.10 x86_64 - Varnish Cache 4.1.
 - Updates pattern for static assets to include web fonts and SVG images and remove SWF.
 - Removes response header that indicate Varnish version.
 - Adds `VARNISH_AUTOSTART_VARNISHD_WRAPPER` for disabling varnishd autostart.
+- Adds `VARNISH_AUTOSTART_VARNISHNCSA_WRAPPER` to enable access logs.
+- Adds `VARNISH_VARNISHNCSA_FORMAT` set the access log format string.
 
 ### 1.4.4 - 2018-06-22